refactor
Внесение небольших коррекировок, согласно комментариям
